refactor(HIV infant testing conditions): Refactor hiv infant testing: Adjust age requirement and relevant conditions
This commit simplifies the HIV infant testing conditions in both infant testing and child predicates. Initially, infant testing was required only under the age of 18 months. This constraint has been removed, so testing covers a broader age range. The test for infant testing not required after 18 months was also removed from the infant testing file. In the child predicates file, the condition 'child_age_in_months <= 18' was taken off. Additionally, "visit_code == '3000'" was added to the governing conditions for when HIV testing is required. This overall change provides a more comprehensive coverage for required tests in the study.