Refine maternal visit filter in caregiver_predicates.py
This commit modifies the method 'currently_pregnant' in the caregiver_predicates.py file. We now also check the 'child_subject_identifier', alongside the 'subject_identifier', to determine whether a caregiver is currently pregnant. This change ensures a more accurate filtering of maternal visits and tackles the issue where multiple visits could be relevant for a particular subject.