(feature): Add childhood lead exposure risk rule function
This update includes the addition of a rule function 'func_childhood_lead_exposure_risk_required' to the child_predicates.py file and a 'childhood_lead_exposure_risk' rule to the child_visit_rules.py file. The function checks if there's a need for a childhood lead exposure risk assessment during follow-up visits, and the rule enforces this condition.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>