(refactor): Update condition in child_predicates return logic
The return logic of the child_predicates.py module has been updated for better functionality. Previously, it returned true only if there was no previous instance and it was a follow-up. Now, the condition checks if the time difference between the current visit and the last instance is over a year in addition to the original conditions. This change allows for more accurate results when checking for follow-up visits after a year.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>