(refactor): Update calculation for earliest appointment date
Updated the calculation for the earliest appointment date in the child predicates file. Instead of directly comparing the appointment datetime with the previous instance report datetime, the calculation now takes into account the visit definition and timezone. This allows for a more accurate determination of time differences, specifically for child appointments scheduled in the 'Africa/Gaborone' timezone.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>