(Feature): Add TB screening and HEU status functions
Two new functions have been added to the child predicates. The `func_child_tb_screening_required` function checks if a child's TB screening is needed based on a series of tests. In addition to the TB screening function, the `func_heu_status_disclosed` has been introduced to determine the HIV exposure of a child whether the HEU status is disclosed or not.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>