(refactor): Update caregiver predicates to handle appointment schedules
The code associated with caregiver predicates has been refactored. The key areas of change include updating how the previous appointments are fetched. Instead of simply fetching the previous appointment, the code now retrieves all previous appointments that match a list of schedule names. Another significant change is the modification to the rules for determining valid child age. The rules now explicitly return if the child age is valid or if it is a follow-up visit.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>